scripts/ansible.Makefile: add run helper
Add run_ansible_playbook_local helper. This allows to centralize
ansible-playbook calls and extend the helper/helpers with extra
functionality.

This change allows to overwrite the KDEVOPS_EXTRA_VARS variable for the
users of run_ansible_playbook_local helper.

If KDEVOPS_EXTRA_VARS is not overwritten by the user when running make,
there are no functional changes.

Note that we don't need to specify anymore ansible_python_interpreter as
it's part of the extra_vars.yml file.

Signed-off-by: Daniel Gomez <da.gomez@samsung.com>