ansible.Makefile: enhance ansible verbosity
Ansible verbosity is controlled via V=2 parameter/value. This however
has a limited configuration as it hardcodes the output of the -v,
--verbose command of ansible-playbook to level 3 (-vvv). Drop support
for V=2 mode in favor of AV=<number> where <number> indicates the level
of verbosity from 0 (empty string) to max 6 (-vvvvvv).

Add a helper script in python to handle the new AV parameter. This makes
it easier to check for corner cases.

Also, restore verbosity help from commit 6539f3f ("Merge commit
'6539f3f6b5cf1393cb63f9ad9aac51877f088734'") and adapt V=2 to the new
AV= parameter.

Add a maintainer entry for this new Ansible wrapper. A reason to add
a specific entry here is to extend the new ansible.Makefile with all
Ansible calls across kdevops and make it easier to maintain in the long
term.

Signed-off-by: Daniel Gomez <da.gomez@samsung.com>