guestfs: bringup: check if domain exists
virsh domstate fails (exit 1) if domain checked does not exist. This
conflicts when debugging the script with set -euxo pipefile. Replace the
command with virsh list --all and search for the domain in the list.
Also, capture the state to avoid trying to start a domain that is
already running.

Signed-off-by: Daniel Gomez <da.gomez@samsung.com>