.b4-config: add b4 support
To be able to support b4 properly, we need to configure a custom
get_maintainers.pl argument list in b4. Use the same b4 default
arguments + "--no-tree" to run the script outside a kernel tree.

Documentation b4.send-auto-to-cmd and b4.send-auto-cc-cmd:
https://b4.docs.kernel.org/en/latest/
config.html#contributor-oriented-settings

This allows to run 'b4 prep --auto-to-cc', to collect the relevant
addresses from the MAINTAINERS file for the commits/files included in
the contributor patch series.

For more information about contributor workflow:
https://b4.docs.kernel.org/en/latest/contributor/
overview.html#what-s-the-b4-contributor-workflow

Signed-off-by: Daniel Gomez <da.gomez@samsung.com>
Reviewed-by: Luis Chamberlain <mcgrof@kernel.org>