Revert "bringup: devconfig: fix python3-apt debian issue"
This reverts commit bcea53d0710cc0722520430277c80279e90e0863

Commit bcea53d0710cc0722520430277c80279e90e0863 ("bringup: devconfig:
fix python3-apt debian issue") was introduced as workaround to fix a
Debian python3-apt issue. We now know this was the actual problem [2]:

"APT has a fluctuating ABI that
depends on which std:: symbols it calls, and which of those calls
the compiler inlines and omits the symbols for, so it can change
with every update to APT or the compiler."

Ensuring distro and especially apt package is uptodate we align always
packages versions to their latest available. Do this at bringup when
image is firstboot. In addition, add a new firstconfig target, to make
sure we have the system upgraded even if the image has already been
booted and we have an outdated system.

[1]
https://github.com/linux-kdevops/linux-mm-kpd/actions/runs/13677761684/job/38242628757

[2]
https://lore.kernel.org/kdevops/20250208134034.GA4079710@debian.org/T/#u

Signed-off-by: Daniel Gomez <da.gomez@samsung.com>