Generate Kconfig environment in Ansible
This is an effort to replace the scripts/* we use from Kconfig to
retrieve any possible Kconfig value (y/n/string) with Ansible. I believe
this has been discussed a few times on the mailing list, perhaps as a
potential idea. This aims to explore whether this approach makes sense
and what are the options.

The new workflow is:
* Run make menuconfig
* menuconfig runs kconfig-env target
* kconfig-env target calls Ansible kconfig.yml playbook
* kconfig.yml playbook generates a Kconfig fragment with the environment
(.env.config)
* .env.config is merged with .config
* make olddefconfig is run to generate a valid .config for menuconfig
* menuconfig is shown with the environment generated variables

While this adds new targets, it also removes the $shell calls in
Kconfig. However, when merging fragments, it appears that hidden symbols
cannot be merged and updated in .config, unless these are visible (and
therefore editable). So for this reason, a new Environment menu has
been created to list all these new vars. Any ideas/suggestion to keep
them invisible?

Note that only CONFIG_TOPDIR_PATH_SHA256SUM= is generated with Ansible
and added/merged to the Kconfig environment menu vars.

Thoughts?

# Describe the purpose of this series. The information you put here
# will be used by the project maintainer to make a decision whether
# your patches should be reviewed, and in what priority order. Please be
# very detailed and link to any relevant discussions or sites that the
# maintainer can review to better understand your proposed changes. If you
# only have a single patch in your series, the contents of the cover
# letter will be appended to the "under-the-cut" portion of the patch.

# Lines starting with # will be removed from the cover letter. You can
# use them to add notes or reminders to yourself. If you want to use
# markdown headers in your cover letter, start the line with ">#".

# You can add trailers to the cover letter. Any email addresses found in
# these trailers will be added to the addresses specified/generated
# during the b4 send stage. You can also run "b4 prep --auto-to-cc" to
# auto-populate the To: and Cc: trailers based on the code being
# modified.

To: Luis Chamberlain <mcgrof@kernel.org>
Cc: kdevops@lists.linux.dev
Cc: Daniel Gomez <da.gomez@kernel.org>
Signed-off-by: Daniel Gomez <da.gomez@samsung.com>

---
Changes in v2:
- EDITME: describe what is new in this series revision.
- EDITME: use bulletpoints and terse descriptions.
- Link to v1: https://lore.kernel.org/r/20250328-kconfig-env-with-ansible-v1-0-3231c0272457@samsung.com



--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 2,
    "change-id": "20250328-kconfig-env-with-ansible-33e3c3b3ae97",
    "prefixes": [
      "RFC"
    ],
    "history": {
      "v1": [
        "20250328-kconfig-env-with-ansible-v1-0-3231c0272457@samsung.com"
      ]
    }
  }
}