Makefile: use inventory from ansible.cfg
We can define the Ansible inventory via ansible.cfg inventory variable.
Make use of it and remove all the -i/--inventory arguments in all the
ansible-playbook calls.

Target selection is now handled in two ways:

At the playbook level (hosts: <targets>): E.g. using patterns like all,
baseline:dev (which targets all nodes except localhost), or localhost.
While all!localhost could be used as an alternative to baseline:dev,
it is not semantically correct for kdevops. This is because it would
prevent playbooks from running on the controller host that also functions
as a target node.

So, most of the playbooks have been converted from all -> baseline:dev.
Except for bootlinux playbook, where all is kept as the playbook runs
limited to localhost in linux-clone-9p and limited to baseline:dev for
the rest of the targets.

- At command line: ansible-playbook --limit 'SUBSET': In order to
control the hosts when the playbook defines something wider (such as the
case where we want to run only on baseline targets for example), we use
the ansible-playbook argument --limit.

Use --limits argument with the proper syntax [1]: with quotes and colons
(for list of targets or group of targets).

Ensure the target string in CONFIG_ANSIBLE_CFG_INVENTORY excludes the
double quotes to avoid the ANSIBLE_CFG_INVENTORY target to be triggered
every time.

Also, include support for KDEVOPS_HOSTS via _SET_BY_CLI to keep
compatibility with the KDEVOPS_HOSTS Makefile variable that could be
overwritten by the user when make was invoked to define a specific hosts
file.

[1]
https://docs.ansible.com/ansible/latest/inventory_guide/intro_patterns.html

Signed-off-by: Daniel Gomez <da.gomez@samsung.com>