Makefile: use inventory from ansible.cfg
We can define the Ansible inventory via ansible.cfg inventory variable.
Make use of it and remove all the -i/--inventory arguments in all the
ansible-playbook calls.

Target control is now handled in 2 different ways:
- At playbook (hosts: <targets>): E.g. all, all!localhost (all targets
except localhost), localhost. An alternative to all!localhost is to
define all the groups such as baseline,dev. The former option was chosen
for brevity.

So, most of the playbooks have been converted from all ->
all!localhosts. Except for bootlinux playbook, where all is kept as
the playbook runs limited to localhost in linux-clone-9p and limited to
all:!localhosts for the rest of the targets.

- At command line: ansible-playbook --limit 'SUBSET'. In order to
control the hosts when the playbook defines something wider (such as the
case where we want to run only on baseline targets for example), we use
the ansible-playbook argument --limit.

Use --limits argument with the proper syntax [1]: with quotes and colons
(for list of targets or group of targets).

[1]
https://docs.ansible.com/ansible/latest/inventory_guide/intro_patterns.html

Signed-off-by: Daniel Gomez <da.gomez@samsung.com>