ansible_cfg: add support to change ansible.cfg file location
This adds support to change the Ansible Configuration filename and
location via Kconfig.

Ansible checks the ansible.cfg file using the following order [1]:

* ANSIBLE_CONFIG (environment variable if set)
* ansible.cfg (in the current directory)
* ~/.ansible.cfg (in the home directory)
* /etc/ansible/ansible.cfg

In order to set a custom path, export the ANSIBLE_CONFIG [2] that points
to the file selected in Kconfig. But only when the Kconfig option is not
empty. In such case, let the user handle the ANSIBLE_CONFIG variable.

[1]
https://docs.ansible.com/ansible/latest/reference_appendices/config.html#the-configuration-file

[2]
https://docs.ansible.com/ansible/latest/reference_appendices/config.html#envvar-ANSIBLE_CONFIG

It also fixes the target to generate the file only if it doesn't exist,
instead of regenerating it every time make is invoked. Removing the
quotes from the variable used as target fixes the problem. Reported
by Chuck.

In addition, fix verbose mode (V=1) for the ansible-playbook command by
removing the extra @.

Remove the file from MAINTAINERS list as it is autogenerated and not
in tree.

Reported-by: Chuck Lever <chuck.lever@oracle.com>
Signed-off-by: Daniel Gomez <da.gomez@samsung.com>