ansible_cfg: add support to change ansible.cfg file location
This adds support to change the Ansible Configuration filename and
location via Kconfig.

It also fixes the target to generate the file only if it doesn't exist,
instead of regenerating it every time make is invoked. Removing the
quotes from the variable used as target fixes the problem.

Remove the file from MAINTAINERS list as it is autogenerated and not
in tree.

Reported-by: Chuck Lever <chuck.lever@oracle.com>
Signed-off-by: Daniel Gomez <da.gomez@samsung.com>