fstests: use ansible_hostname for configuration
The fstest configuration file uses format <hostname>.config. In order
to generate the correct name for baremetal, use the ansible_hostname
variable instead of the ansible_host. The former uses the actual
hostname of the target while the latter uses the name on the inventory
file. This allows to generate the appropiate filename for both type
of workflows.

Signed-off-by: Daniel Gomez <da.gomez@samsung.com>