Add configuration file for cli
We use GNU Make as a wrapper mainly for Kconfig and Ansible. In order
to assist kdevops users to change certain options we expose certain
features to the cli interface via Makefile parameters. A few examples
are the verbosity of the ansible-playbook command (-v, -vv...) via AV
parameter. Or the Linux tree git reference (Kconfig) to be used via
LINUX_TREE_REF. This seems convenient and has been increasing over time.
Enough parameters have been added than I personally tend to forget
the exact variable names. For that, documentation may be best but when
using a large combination of these options, the cli becomes inconvenient
by itself. To help with command cli parameters, this patch adds a
file params.mk.sample that will generate a params.mk (if it doesn't
exist already) with all parameters disabled. But the file will contain
examples of how to use each of the options available via cli interface.
The main Makefile will parse this first and will make it available as
if the user would pass them through the command line making it easier to
handle, specially when multiple options are used.

Thoughts?

# Describe the purpose of this series. The information you put here
# will be used by the project maintainer to make a decision whether
# your patches should be reviewed, and in what priority order. Please be
# very detailed and link to any relevant discussions or sites that the
# maintainer can review to better understand your proposed changes. If you
# only have a single patch in your series, the contents of the cover
# letter will be appended to the "under-the-cut" portion of the patch.

# Lines starting with # will be removed from the cover letter. You can
# use them to add notes or reminders to yourself. If you want to use
# markdown headers in your cover letter, start the line with ">#".

# You can add trailers to the cover letter. Any email addresses found in
# these trailers will be added to the addresses specified/generated
# during the b4 send stage. You can also run "b4 prep --auto-to-cc" to
# auto-populate the To: and Cc: trailers based on the code being
# modified.

To: Luis Chamberlain <mcgrof@kernel.org>
To: Chuck Lever <chuck.lever@oracle.com>
Cc: kdevops@lists.linux.dev
Signed-off-by: Daniel Gomez <da.gomez@samsung.com>

--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 1,
    "change-id": "20250703-params-mk-e7814194dc26",
    "prefixes": [],
    "prerequisites": []
  }
}