sysbench: decouple DUT name when A/B testing is enabled
When A/B testing is enabled (default in sysbench) and a physical device
is assigned to one guest (e.g., via PCIe passthrough to the baseline
guest), the sysbench dev guest run may fail because it attempts to
access a device that isn't present on that host.

This change decouples the DUT name between baseline and dev guests.
Note that in scenarios with multiple dev guests, further decoupling of
the variable would be required, similar to how PCIe passthrough handles
device-to-host assignment. Stop decoupling at group level as I don't
have a particular use case yet for that scenario.

# Describe the purpose of this series. The information you put here
# will be used by the project maintainer to make a decision whether
# your patches should be reviewed, and in what priority order. Please be
# very detailed and link to any relevant discussions or sites that the
# maintainer can review to better understand your proposed changes. If you
# only have a single patch in your series, the contents of the cover
# letter will be appended to the "under-the-cut" portion of the patch.

# Lines starting with # will be removed from the cover letter. You can
# use them to add notes or reminders to yourself. If you want to use
# markdown headers in your cover letter, start the line with ">#".

# You can add trailers to the cover letter. Any email addresses found in
# these trailers will be added to the addresses specified/generated
# during the b4 send stage. You can also run "b4 prep --auto-to-cc" to
# auto-populate the To: and Cc: trailers based on the code being
# modified.

To: Luis Chamberlain <mcgrof@kernel.org>
Cc: kdevops@lists.linux.dev
Signed-off-by: Daniel Gomez <da.gomez@samsung.com>

--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 1,
    "change-id": "20250916-sysbench-ab-dut-d29b07bb37b5",
    "prefixes": []
  }
}